virtualCRUD/requirements.txt
requests
django
pymysql
djangorestframework
django-import-export
geopy